refactor: extract package librelingo_types
