+++ apps/librelingo_tools/librelingo_tools/__init__.py
@@ -2,7 +2,6 @@
 Export LibreLingo courses in the JSON format expected by the web app
 """
 
-import librelingo_tools.data_types
 import librelingo_tools.export
 import librelingo_tools.yaml_loader
