+++ apps/librelingo_tools/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@librelingo/tools",
+  "private": true,
+  "scripts": {
+    "test": "poetry run ptw"
+  }
+}
