+++ apps/librelingo_types/librelingo_types/__init__.py
@@ -1,0 +1,7 @@
+"""
+Data types to be used in Python packages for LibreLingo
+"""
+
+from librelingo_types.data_types import *
+
+__version__ = '0.1.1'
